warehouse/asyncworker/pyproject.toml
[project]
name = "asyncworker"
version = "0.1.0"
description = "Async workers for OSO"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-pubsub>=2.33.0",
    "pydantic-cli>=10.0.0",
    "pydantic-settings>=2.11.0",
    "query-rewriter",
    "osoprotobufs",
    "oso-core",
    "oso-dagster",
]

[tool.uv.sources]
query-rewriter = { workspace = true }
pyoso = { workspace = true }
osoprotobufs = { workspace = true }
oso-core = { workspace = true }
oso-dagster = { workspace = true }

[project.scripts]
asyncworker = "asyncworker.cli:cli"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.uv.build-backend]
module-name = "asyncworker"
module-root = ""
